+++ classes/Assets.php
@@ -16,10 +16,6 @@ use Grav\Common\Grav;
 use Grav\Common\Utils;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Filesystem\Exception\IOExceptionInterface;
-// use Windwalker\Http\HttpClient;
-use GuzzleHttp\Client;
-// use GuzzleHttp\Psr7\Request;
-use Grav\Framework\Psr7\Request;
 
 /**
  * Assets Builder
@@ -41,12 +37,13 @@ class Assets
      * @param Filesystem $Filesystem Instance of Symfony\Component\Filesystem\Filesystem.
      * @param Timer      $Timer      Instance of Grav\Plugin\StaticGenerator\Timer.
      */
-    public function __construct(Filesystem $Filesystem, Timer $Timer)
+    public function __construct(Filesystem $Filesystem, Timer $Timer, bool $offline = false)
     {
         $this->streams = array();
         $this->schemes = array();
         $this->Filesystem = $Filesystem;
         $this->Timer = $Timer;
+        $this->offline = $offline;
         $this->grav = Grav::instance();
         foreach ($this->grav['locator']->getSchemes() as $stream) {
             $this->schemes[$stream] = Utils::url($stream . '://');
@@ -67,80 +64,32 @@ class Assets
      */
     public function copy(string $asset, string $location, bool $force): array
     {
-        // dump($asset);
-        if (Utils::startsWith($asset, '/user')) {
-            // dump('user');
-        } elseif (Utils::startsWith($asset, '/system')) {
-            // dump('system');
-        } else {
-            dump($asset);
-            // dump(basename($asset));
-            // dump(dirname($asset));
-            // dump(parse_url($asset));
-            // dump($location);
-            $url = parse_url($asset);
-            $target = $location . DS . 'assets' . DS . $url['host'] . $url['path'];
-            dump($target);
-            // file_put_contents($target, fopen($asset, 'r'));
-            $this->Filesystem->copy($asset, $target);
-            // new Request(
-            //     'GET',
-            //     $asset,
-            //     [
-            //         'connect_timeout' => 2,
-            //         'sink' => $target,
-            //         'timeout' => 3,
-            //     ]
-            // );
-            /* $Request = new Request('GET', $asset);
-            dump($Request);
-            dump($Request->getBody());
-            dump($Request->getBody()->getContents());
-            // dump($Request->getBody()->read(0));
-            dump($Request->getBody()->__toString());
-            dump($Request->getBody()->message);
-            dump($Request->getBody()->message->getBody()); */
-            
-            // include __DIR__ . '/../vendor/autoload.php';
-            // $file_path = fopen($target, 'w');
-            // dump($file_path);
-            // $client = new Client();
-            // $response = $client->get($asset, ['sink' => $file_path]);
-            // dump($response);
-
-            // $fileHandle = fopen($target, "wb");
-            // dump($fileHandle);
-            // try {
-            //     $client = new Client();
-            //     $response = $client->get($asset, [
-            //         'sink' => $fileHandle
-            //     ]);
-            // } catch (RequestException $e) {
-            //     throw new ReportFileDownloadException(
-            //         "Can't download report file $asset"
-            //     );
-            // } finally {
-            //     @fclose($fileHandle);
-            // }
-
-            // dump('other');
-            // $http = new HttpClient;
-            // $dest = '/path/to/local/file.zip';
-            // $response = $http->download('http://example.com/file.zip', $dest);
-        }
-        return [];
         if (empty($asset)) {
             return [];
         }
         $location = $location . DS . 'assets';
+        if (Utils::startsWith($asset, '/user')) {
+            $source = GRAV_ROOT . $asset;
+            $target = $location . $asset;
+        } elseif (Utils::startsWith($asset, '/system')) {
+            $source = GRAV_ROOT . $asset;
+            $target = $location . $asset;
+        } else {
+            if ($this->offline === true) {
+                return [];
+            }
+            $source = $asset;
+            $url = parse_url($asset);
+            $target = $location . DS . $url['host'] . $url['path'];
+        }
         try {
             if ($force) {
-                $this->Filesystem->remove($location . $asset);
+                $this->Filesystem->remove($target);
             }
-            $this->Filesystem->copy(GRAV_ROOT . $asset, $location . $asset);
+            $this->Filesystem->copy($source, $target);
             return [
                 'item' => basename($asset),
-                'location' => $location . $asset,
+                'location' => $target,
                 'time' => Timer::format($this->Timer->getTime())
             ];
         } catch (\Exception $e) {